(홍진표)부서일정 삭제 구현(부서장만 삭제가능)

+++ src/main/java/com/goodee/groupware/sevice/ScheduleService.java
@@ -141,28 +141,35 @@ public class ScheduleService {
 		cnt = scheduleMapper.getMemberLevelCount(paramMap);
 		log.debug("\u001B[31m"+"ScheduleService.addDepartmentSchedule() cnt : "+cnt+"\u001B[0m");
 		if(cnt > 0) { // 부서장이 맞으면
-			// 일정 추가
-			row = scheduleMapper.addSchedule(schedule); 
+			row = scheduleMapper.addSchedule(schedule); // 일정 추가
 		} else { // 부서장이 아니면
 			row=0;
 		}
 		log.debug("\u001B[31m"+"ScheduleService.addDepartmentSchedule() row : "+row+"\u001B[0m");
-		
 		return row; 
 	}
 	
-	// 개인 일정 삭제
-	public int deletePersonalSchedule(Schedule schedule) {
+	// 부서 일정 삭제
+	public int deleteDepartmentSchedule(Schedule schedule) {
+		int cnt = 0;
 		int row = 0;
-		// 일정 삭제
-		row = scheduleMapper.deleteSchedule(schedule);
+		String memberLevel = "3부서장";
 		
-		log.debug("\u001B[31m"+"ScheduleService.deletePersonalSchedule() row : "+row+"\u001B[0m");
+		// 매개변수 값들을 Map에 담음
+		Map<String, Object> paramMap = new HashMap<String, Object>();
+		paramMap.put("memberLevel", memberLevel);
+		paramMap.put("memberId", schedule.getMemberId());
+		
+		// 부서장 유무 조회
+		cnt = scheduleMapper.getMemberLevelCount(paramMap);
+		log.debug("\u001B[31m"+"ScheduleService.deleteDepartmentSchedule() cnt : "+cnt+"\u001B[0m");
+		if(cnt > 0) { // 부서장이 맞으면
+		row = scheduleMapper.deleteSchedule(schedule);
+		} else { // 부서장이 아니면
+			row=0;
+		}
+		log.debug("\u001B[31m"+"ScheduleService.deleteDepartmentSchedule() row : "+row+"\u001B[0m");
 		return row;
 	}
-	
-	
-	
-	// 부서 일정 삭제(부서장만 삭제 가능)
 	
 }
